DOCS: Entity 위치 변경에 따른 Refactoring
