+++ src/main/java/spot/spot/domain/job/command/repository/jpa/CertificationRepository.java
@@ -1,8 +1,8 @@
-package spot.spot.domain.job.repository.jpa;
+package spot.spot.domain.job.command.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import spot.spot.domain.job.entity.Certification;
+import spot.spot.domain.job.command.entity.Certification;
 
 @Repository
 public interface CertificationRepository extends JpaRepository<Certification, Long> {
