+++ src/main/java/spot/spot/domain/job/command/repository/jpa/HashTagRepository.java
@@ -1,8 +1,8 @@
-package spot.spot.domain.job.repository.jpa;
+package spot.spot.domain.job.command.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import spot.spot.domain.job.entity.HashTag;
+import spot.spot.domain.job.command.entity.HashTag;
 
 @Repository
 public interface HashTagRepository extends JpaRepository<HashTag, Long> {
