+++ src/main/java/spot/spot/domain/pay/entity/KlayAboutJob.java
@@ -2,7 +2,7 @@ package spot.spot.domain.pay.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import spot.spot.domain.job.entity.Job;
+import spot.spot.domain.job.command.entity.Job;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
