+++ src/main/java/spot/spot/domain/pay/repository/KlayAboutJobRepository.java
@@ -1,7 +1,7 @@
 package spot.spot.domain.pay.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import spot.spot.domain.job.entity.Job;
+import spot.spot.domain.job.command.entity.Job;
 import spot.spot.domain.pay.entity.KlayAboutJob;
 
 import java.util.Optional;
