+++ src/test/java/spot/spot/domain/pay/repository/KlayAboutJobRepositoryTest.java
@@ -4,12 +4,11 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
-import spot.spot.domain.job.entity.Job;
-import spot.spot.domain.job.repository.jpa.JobRepository;
+import spot.spot.domain.job.command.entity.Job;
+import spot.spot.domain.job.query.repository.jpa.JobRepository;
 import spot.spot.domain.pay.entity.KlayAboutJob;
 import spot.spot.domain.pay.entity.PayHistory;
 import spot.spot.domain.pay.entity.PayStatus;
